feat: creates submit button component and fixes all forms
