+++ resources/js/Components/Modal.tsx
@@ -1,68 +1,27 @@
-import { Fragment, PropsWithChildren } from "react";
-import { Dialog, Transition } from "@headlessui/react";
+import { ReactNode, forwardRef } from "react";
 
-export default function Modal({
-    children,
-    show = false,
-    maxWidth = "2xl",
-    closeable = true,
-    onClose = () => {},
-}: PropsWithChildren<{
-    show: boolean;
-    maxWidth?: "sm" | "md" | "lg" | "xl" | "2xl";
-    closeable?: boolean;
-    onClose: CallableFunction;
-}>) {
-    const close = () => {
-        if (closeable) {
-            onClose();
-        }
-    };
-
-    const maxWidthClass = {
-        sm: "sm:max-w-sm",
-        md: "sm:max-w-md",
-        lg: "sm:max-w-lg",
-        xl: "sm:max-w-xl",
-        "2xl": "sm:max-w-2xl",
-    }[maxWidth];
-
-    return (
-        <Transition show={show} as={Fragment} leave="duration-200">
-            <Dialog
-                as="div"
-                id="modal"
-                className="fixed inset-0 flex overflow-y-auto px-4 py-6 sm:px-0 items-center z-50 transform transition-all"
-                onClose={close}
-            >
-                <Transition.Child
-                    as={Fragment}
-                    enter="ease-out duration-300"
-                    enterFrom="opacity-0"
-                    enterTo="opacity-100"
-                    leave="ease-in duration-200"
-                    leaveFrom="opacity-100"
-                    leaveTo="opacity-0"
-                >
-                    <div className="absolute inset-0 bg-gray-500/75 dark:bg-gray-900/75" />
-                </Transition.Child>
-
-                <Transition.Child
-                    as={Fragment}
-                    enter="ease-out duration-300"
-                    enterFrom="opacity-0 translate-y-4 sm:translate-y-0 sm:scale-95"
-                    enterTo="opacity-100 translate-y-0 sm:scale-100"
-                    leave="ease-in duration-200"
-                    leaveFrom="opacity-100 translate-y-0 sm:scale-100"
-                    leaveTo="opacity-0 translate-y-4 sm:translate-y-0 sm:scale-95"
-                >
-                    <Dialog.Panel
-                        className={`mb-6 bg-white dark:bg-gray-800 rounded-lg overflow-hidden shadow-xl transform transition-all sm:w-full sm:mx-auto ${maxWidthClass}`}
-                    >
-                        {children}
-                    </Dialog.Panel>
-                </Transition.Child>
-            </Dialog>
-        </Transition>
-    );
+export interface ModalProps {
+    type: "Edit" | "Add" | "Delete";
+    item: string;
+    form: ReactNode | null;
 }
+
+const Modal = forwardRef<HTMLDialogElement, ModalProps>(function Modal(
+    { type, item, form },
+    ref,
+) {
+    return (
+        <dialog ref={ref} className="modal">
+            <div className="modal-box">
+                <div className="flex flex-col">
+                    <span className="text-3xl font-semibold pb-6 text-neutral">
+                        {type} {item}
+                    </span>
+                    <div>{form}</div>
+                </div>
+            </div>
+        </dialog>
+    );
+});
+
+export default Modal;
