+++ resources/js/Components/forms/inputs/SubmitButton.tsx
@@ -1,0 +1,8 @@
+export function SubmitButton({ errorMessage }: { errorMessage: string }) {
+    return (
+        <label className="form-control pt-4">
+            <input className="btn btn-primary" type="submit" />
+            <div className="text-error text-center pt-2">{errorMessage}</div>
+        </label>
+    );
+}
