+++ resources/js/Components/forms/inputs/index.js
@@ -2,3 +2,4 @@ export { CloseX } from "./CloseX";
 export { TextAreaInput } from "./TextAreaInput";
 export { DropdownInput } from "./DropdownInput";
 export { TextInput } from "./TextInput";
+export { SubmitButton } from "./SubmitButton";
