+++ resources/js/Pages/Profile/Partials/DeleteUserForm.tsx
@@ -2,7 +2,7 @@ import { useRef, useState, FormEventHandler } from "react";
 import DangerButton from "@/Components/DangerButton";
 import InputError from "@/Components/InputError";
 import InputLabel from "@/Components/InputLabel";
-import Modal from "@/Components/Modal";
+import Modal from "@/Components/Modal-Old";
 import SecondaryButton from "@/Components/SecondaryButton";
 import TextInput from "@/Components/TextInput";
 import { useForm } from "@inertiajs/react";
